Renamed sample test.
